complete default package installed upon shell provisionning

+++ tests/vagrant/puppet_modules_setup.rb
@@ -2,11 +2,11 @@
 ##########################################################################
 # puppet_module_setup.rb
 # @author Sebastien Varrette <Sebastien.Varrette@uni.lu>
-# Time-stamp: <Tue 2019-01-29 08:34 svarrette>
+# Time-stamp: <Thu 2019-01-31 20:05 svarrette>
 #
 # @description Prepare the Vagrant box to test this Puppet module
 #
-# Copyright (c) 2014-2017 Sebastien Varrette <Sebastien.Varrette@uni.lu>
+# Copyright (c) 2014-2019 Sebastien Varrette <Sebastien.Varrette@uni.lu>
 # .             http://varrette.gforge.uni.lu
 ##############################################################################
 
@@ -19,51 +19,32 @@ include FalkorLib::Common
 # Load metadata
 basedir   = File.directory?('/vagrant') ? '/vagrant' : Dir.pwd
 jsonfile  = File.join(basedir, 'metadata.json')
-# puppetdir = '/etc/puppetlabs'
+puppetdir = File.join(basedir, 'tests', 'vagrant', 'puppet')
 
 error 'Unable to find the metadata.json' unless File.exist?(jsonfile)
 
 metadata   = JSON.parse(IO.read(jsonfile))
 name       = metadata['name'].gsub(%r{^[^\/-]+[\/-]}, '')
-modulepath = `puppet config print modulepath`.chomp
-moduledir  = modulepath.split(':').first
+#modulepath = `puppet config print modulepath`.chomp
+#moduledir  = modulepath.split(':').first
+moduledir = File.join(puppetdir, 'modules')
 
-run %( cd #{moduledir}/.. && librarian-puppet clean && rm Puppetfile* )
-run %( ln -s /vagrant/metadata.json #{moduledir}/../ )
-run %( cd #{moduledir}/.. && librarian-puppet install --verbose )
-
+unless File.directory?(File.join(moduledir, 'stdlib'))
+  run %( cd #{moduledir}/.. && librarian-puppet clean && rm Puppetfile* )
+  run %( ln -s /vagrant/metadata.json #{moduledir}/../ )
+  run %( cd #{moduledir}/.. && librarian-puppet install --verbose )
+end
 # metadata["dependencies"].each do |dep|
-# 	lib = dep["name"]
+#   lib = dep["name"]
 #     shortname = lib.gsub(/^.*[\/-]/,'')
 #     action = File.directory?("#{moduledir}/#{shortname}") ? 'upgrade --force' : 'install'
-# 	run %{ puppet module #{action} #{lib} }
+#   run %{ puppet module #{action} #{lib} }
 # end
 
-puts "Module path: #{modulepath}"
+#puts "Module path: #{modulepath}"
 puts "Moduledir:   #{moduledir}"
 
-info "set symlink to the '#{basedir}' module for local developments"
-run %( ln -s #{basedir} #{moduledir}/#{name} ) unless File.exist?("#{moduledir}/#{name}")
-
-# Use of 'hiera.yaml' version 3 is deprecated. It should be converted to version 5
-# hiera = '/etc/puppetlabs/puppet/hiera.yaml'
-
-# Prepare hiera
-# unless File.exists?('/etc/puppet/hiera.yaml')
-#     run %{ ln -s /etc/hiera.yaml /etc/puppet/hiera.yaml } if File.exists?("/etc/hiera.yaml")
-# end
-# # # hieracfg = YAML::load_file('/etc/hiera.yaml')
-# # # [ '/vagrant/tests/hiera' ].each do |d|
-# # #     hieracfg[:datadir] = [] if hieracfg[:datadir].nil?
-# # #     hieracfg[:datadir] << d #if File.directory?('#{d}')
-# # # end
-# # # hieracfg[:hierarchy] = [] if hieracfg[:hierarchy].nil?
-# # # hieracfg[:hierarchy] << 'common' unless hieracfg[:hierarchy].include?('common')
-# hieracfg = {
-#     :backends  => [ 'yaml' ],
-#     :hierarchy => [ 'defaults', 'common' ],
-#     :yaml      => {
-#         :datadir   => '/vagrant/tests/hiera',
-#     }
-# }
-# FalkorLib::Common.store_config('/etc/hiera.yaml', hieracfg, {:no_interaction => true})
+unless File.exist?("#{moduledir}/#{name}")
+  info "set symlink to the '#{basedir}' module for local developments"
+  run %( ln -s #{basedir} #{moduledir}/#{name} )
+end
